feat: api model
