+++ src/web/src/@types/index.ts
@@ -1,4 +1,3 @@
 export * from './api';
 export * from './direction';
-export * from './models';
 export * from './styles';
